Rename novelty_interface module to ond_adapter
