+++ sailon/__init__.py
@@ -1,2 +1,3 @@
+from .ond_adapter import ONDAdapter
 from .random_novelty_detector import RandomNoveltyDetector
 from .dummy_interface import DummyInterface
